--- src/main/java/com/pickle/pickle/domain/dto/UserDto.java
@@ -1,4 +1,0 @@
-package com.pickle.pickle.domain.dto;
-
-public class UserDto {
-}
