+++ src/main/java/com/pickle/server/PickleApplication.java
@@ -1,4 +1,4 @@
-package com.pickle.pickle;
+package com.pickle.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
