+++ src/main/java/com/pickle/server/common/util/response/BaseApiResponse.java
@@ -1,4 +1,4 @@
-package com.pickle.pickle.util.response;
+package com.pickle.server.common.util.response;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.*;
