+++ src/main/java/com/pickle/server/user/domain/User.java
@@ -1,13 +1,17 @@
 
-package com.pickle.pickle.domain.entity;
+package com.pickle.server.user.domain;
+
+
+
+import com.pickle.server.common.Timestamped;
+import lombok.Getter;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
-
+@Getter
 @Entity
-@Table(name = "user")
 public class User extends Timestamped {
 
     @Id
