+++ src/main/java/com/pickle/server/user/dto/UserDto.java
@@ -1,0 +1,4 @@
+package com.pickle.server.user.dto;
+
+public class UserDto {
+}
