+++ src/main/java/com/pickle/server/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.pickle.pickle.repository;
+package com.pickle.server.user.repository;
 
-import com.pickle.pickle.domain.entity.User;
+import com.pickle.server.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
