+++ src/test/java/com/pickle/server/PickleApplicationTests.java
@@ -1,4 +1,4 @@
-package com.pickle.pickle;
+package com.pickle.server;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
